using modal button to submit form
