+++ src/main/resources/i18n/messages.properties
@@ -2549,40 +2549,40 @@ AdminUsersTable.created=Date Created
 CreateNewUser.title=Add User
 CreateNewUser.button=Add User
 
-CreateNewUserForm.form.username.label=Username
-CreateNewUserForm.form.username.required=A username is required.
-CreateNewUserForm.form.username.min=Must be at least 3 characters long.
-CreateNewUserForm.form.firstName.label=First Name
-CreateNewUserForm.form.firstName.required=A first name is required.
-CreateNewUserForm.form.firstName.min=Must be at least 2 characters long.
-CreateNewUserForm.form.lastName.label=Last Name
-CreateNewUserForm.form.lastName.required=A last name is required.
-CreateNewUserForm.form.lastName.min=Must be at least 2 characters long.
-CreateNewUserForm.form.email.label=Email
-CreateNewUserForm.form.email.required=An email is required.
-CreateNewUserForm.form.email.type=Must be a valid email.
-CreateNewUserForm.form.email.min=Must be at least 5 characters long.
-CreateNewUserForm.form.phoneNumber.label=Phone Number
-CreateNewUserForm.form.phoneNumber.min=Must be at least 4 characters long.
-CreateNewUserForm.form.locale.label=Language
-CreateNewUserForm.form.locale.required=A language is required.
-CreateNewUserForm.form.role.label=Role
-CreateNewUserForm.form.role.required=A role is required.
-CreateNewUserForm.form.activate.label=Send User Activation Email
-CreateNewUserForm.form.label.password=Password
-CreateNewUserForm.form.button.submit=Create User
-CreateNewUserForm.emailConfigured.alert.title=Missing Email Configuration
-CreateNewUserForm.emailConfigured.alert.description=User activation emails are disabled due to this instance of IRIDA not having a configured mail server. Please enter a password for the new user account.
-CreateNewUserForm.changePassword.alert.title=Password Policy
-CreateNewUserForm.changePassword.alert.description=Please choose a strong password that meets the following criteria:
-CreateNewUserForm.changePassword.alert.rule1=Must not be empty.
-CreateNewUserForm.changePassword.alert.rule2=Must be at least 8 characters long.
-CreateNewUserForm.changePassword.alert.rule3=Must contain at least 1 uppercase letter.
-CreateNewUserForm.changePassword.alert.rule4=Must contain at least 1 lowercase letter.
-CreateNewUserForm.changePassword.alert.rule5=Must contain at least 1 number.
-CreateNewUserForm.changePassword.alert.rule6=Must contain at least 1 special character. (eg. !@#$%^&*()+?/<>=.\\{})
-CreateNewUserForm.notification.success=New user created successfully.
-CreateNewUserForm.notification.error=An error has occurred while creating the new user. Please check all the fields.
+CreateNewUser.form.username.label=Username
+CreateNewUser.form.username.required=A username is required.
+CreateNewUser.form.username.min=Must be at least 3 characters long.
+CreateNewUser.form.firstName.label=First Name
+CreateNewUser.form.firstName.required=A first name is required.
+CreateNewUser.form.firstName.min=Must be at least 2 characters long.
+CreateNewUser.form.lastName.label=Last Name
+CreateNewUser.form.lastName.required=A last name is required.
+CreateNewUser.form.lastName.min=Must be at least 2 characters long.
+CreateNewUser.form.email.label=Email
+CreateNewUser.form.email.required=An email is required.
+CreateNewUser.form.email.type=Must be a valid email.
+CreateNewUser.form.email.min=Must be at least 5 characters long.
+CreateNewUser.form.phoneNumber.label=Phone Number
+CreateNewUser.form.phoneNumber.min=Must be at least 4 characters long.
+CreateNewUser.form.locale.label=Language
+CreateNewUser.form.locale.required=A language is required.
+CreateNewUser.form.role.label=Role
+CreateNewUser.form.role.required=A role is required.
+CreateNewUser.form.activate.label=Send User Activation Email
+CreateNewUser.form.label.password=Password
+CreateNewUser.form.button.submit=Create User
+CreateNewUser.emailConfigured.alert.title=Missing Email Configuration
+CreateNewUser.emailConfigured.alert.description=User activation emails are disabled due to this instance of IRIDA not having a configured mail server. Please enter a password for the new user account.
+CreateNewUser.changePassword.alert.title=Password Policy
+CreateNewUser.changePassword.alert.description=Please choose a strong password that meets the following criteria:
+CreateNewUser.changePassword.alert.rule1=Must not be empty.
+CreateNewUser.changePassword.alert.rule2=Must be at least 8 characters long.
+CreateNewUser.changePassword.alert.rule3=Must contain at least 1 uppercase letter.
+CreateNewUser.changePassword.alert.rule4=Must contain at least 1 lowercase letter.
+CreateNewUser.changePassword.alert.rule5=Must contain at least 1 number.
+CreateNewUser.changePassword.alert.rule6=Must contain at least 1 special character. (eg. !@#$%^&*()+?/<>=.\\{})
+CreateNewUser.notification.success=New user created successfully.
+CreateNewUser.notification.error=An error has occurred while creating the new user. Please check all the fields.
 
 # These are legacy purposes.  Still used on users create and edit pages
 users.id=ID
